adding ChainerASRInterface
