+++ espnet/nets/asr_interface.py
@@ -123,21 +123,6 @@ class ASRInterface:
         """
         raise NotImplementedError("decoders method is not implemented")
 
-    @staticmethod
-    def custom_converter(*args, **kw):
-        """Get customconverter of the model (Chainer only)."""
-        raise NotImplementedError("custom converter method is not implemented")
-
-    @staticmethod
-    def custom_updater(*args, **kw):
-        """Get custom_updater of the model (Chainer only)."""
-        raise NotImplementedError("custom updater method is not implemented")
-
-    @staticmethod
-    def custom_parallel_updater(*args, **kw):
-        """Get custom_parallel_updater of the model (Chainer only)."""
-        raise NotImplementedError("custom parallel updater method is not implemented")
-
 
 predefined_asr = {
     "pytorch": {
